Delete bookmarks from the database when deleting a book
Not sure how I missed this, but the bookmarks where being left in the
database when the books were deleted. This brings it in line with what
happens when using the device to delete a book.

Fixes #1401502 [Delete bookmarks from the database when deleting a book](https://bugs.launchpad.net/calibre/+bug/1401502)

(cherry picked from commit 6320ea10a08332472e38fa873ad98499932bca47)