Bump Kobo firmware version for sneak release of 3.16.9
Bug #1477999  is a report of firmware 3.16.9. This was unexpected, but,
a release to match the release of the Glo HD in Japan is expected. These
might be related, but there is probably another version. So, bumping to
3.16.11. Fixes #1477999 [New Driver 3.16.9 for Kobo Glo HD isn't supported yet](https://bugs.launchpad.net/calibre/+bug/1477999)

(cherry picked from commit 53ef5a7c2a75eb4ab5254c12bcd26b8fcd0d05d7)