calibre-uninstall: use the calibre-debug executable as the interpreter
Similar to the problem with Ubuntu not having a python command for the
linux-installer.py to run, we also don't have a python command for the
uninstall script to run.

In this case, we don't need painful shellscript since we can run the
uninstaller using calibre-debug, which is guaranteed to be a python
interpreter because we installed it. Detect its location across frozen
and non-frozen builds, and let the uninstall script use that for a
shebang.